% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_4_filter.R
\name{mz_filter_magic}
\alias{mz_filter_magic}
\title{MZ-OBJ Filter-Magic}
\usage{
mz_filter_magic(input_mz_obj, min_intensity, min_missingness = FALSE)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{min_intensity}{\cr
Int       : Threshold to blank data under. (as a intesity value)\cr
: Default values: (Pos=10000, Neg=5000)}

\item{min_missingness}{\cr
Float     : Threshold to blank data under. (as a decimal 10\% = 0.1)

Dependencies : dplyr}
}
\value{
Returns an MZ-OBJ
}
\description{
Apply all filters as suggested by SumR.
}
