% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_4_tagging.R
\name{mz_isotope_hunter}
\alias{mz_isotope_hunter}
\title{MZ-OBJ IsotopeTagging}
\usage{
mz_isotope_hunter(
  input_mz_obj,
  iso_target = 1.0034,
  iso_iter = 5,
  tol_ppm = 5e-06,
  cores = 4
)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{iso_target}{\cr
Float : Dalton diff of isotope to hunt}

\item{iso_iter}{\cr
Integer     : How many 'iterations' of iso_target should we look}

\item{tol_ppm}{\cr
Float   : Overwrite csv tolerance, with single value ppm}

\item{cores}{\cr
Integer : How many threads could a multi-threader thread, if a multi-threader could thread threads?

Dependencies : pbapply, dplyr}
}
\value{
Returns a List of MZ that are isotopes of each other
}
\description{
Hunts for isotopes. Defaults to Carbon13
You may use the return to compare intensities of isotopes.
input_mz_obj[input_mz_obj$mz \%in\% output[\link{'100.12345'}],]
}
