% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_3_mass_defect.R
\name{mz_mass_defect}
\alias{mz_mass_defect}
\title{MZ-OBJ MassDefect}
\usage{
mz_mass_defect(
  input_mz_obj,
  plot = TRUE,
  magicNumber1 = 0.00112,
  magicNumber2 = 0.01953
)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{plot}{\cr
Boolean   : To plot or not to plot.}

\item{magicNumber1}{\cr
Float     : MagicNumber1 in "MN1 * data + MN2"}

\item{magicNumber2}{\cr
Float     : MagicNumber2 in "MN1 * data + MN2"

Dependencies : ggplot2, dplyr}
}
\value{
Returns an MZ-OBJ
}
\description{
Mass Defect Calculation.\cr
Zeros any intensities that are lower than MD-Calc.
}
