% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_7_post_processing.R
\name{mz_pp_pivot_longer}
\alias{mz_pp_pivot_longer}
\title{MZ-OBJ Pivot Longer}
\usage{
mz_pp_pivot_longer(input_mz_obj, plot = TRUE)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{plot}{\cr
Boolean   : to plot or not to plot}
}
\value{
MZLONG-OBJ
}
\description{
Not really sure what this does
\itemize{
\item Requires: ggplot2, tidyr
}
}
