% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_6_analysis.R
\name{mz_welch}
\alias{mz_welch}
\title{MZ-OBJ T-Test}
\usage{
mz_welch(input_mz_obj, metadata, phenotype_a, phenotype_b)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{metadata}{\cr
DataFrame: MZ_metadata}

\item{phenotype_a}{\cr
DataFrame: Filtered metadata for phenotypeA}

\item{phenotype_b}{\cr
DataFrame: Filtered metadata for phenotypeB}
}
\value{
DataFrame with columns:
\itemize{
\item rowname: Float: MZ
\item p      : Float: p-value
\item p_05   : Boolean: is P < 0.05
\item p_10   : Boolean: is P < 0.10
\item p_15   : Boolean: is P < 0.15
}
}
\description{
Welch T-test, input should be an log2 of mz_obj
\itemize{
\item Requires: dplyr
'
}
}
