% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_5_post_processing.R
\name{mzl_pp_log_transform}
\alias{mzl_pp_log_transform}
\title{MZLONG-OBJ Log Transform}
\usage{
mzl_pp_log_transform(input_mzlong, plot = TRUE)
}
\arguments{
\item{plot}{\cr
Boolean   : to plot or not to plot}

\item{input_mz_obj}{\cr
DataFrame : Input MZLONG-Obj}
}
\description{
Convert intensities of MZLong to log2
For normal "mz_obj" use log2(mz_obj)
\itemize{
\item Requires: ggplot2
}
}
