% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_1_extract.R
\name{mzml_extract}
\alias{mzml_extract}
\title{MZML Extract into mz-T(time) OBJ
mzT OBJ= dataframe with an mz column, and a column for each scan time's intensity.}
\usage{
mzml_extract(input_mzml)
}
\arguments{
\item{input_mzml}{\cr
String : Filename}
}
\value{
\link{mzT-pos, mzT-neg}
}
\description{
Extracts MZML file, and splits them into pos/neg scan lists
\itemize{
\item Requires: Bioc::mzR
}
}
