% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_1_extract_centroid.R
\name{mzml_extract_magic}
\alias{mzml_extract_magic}
\title{Create MzObj from directory or file}
\usage{
mzml_extract_magic(files = getwd(), cores = 1, ...)
}
\arguments{
\item{files}{\cr
String: File or directory of mzmL files}

\item{cores}{Integer: Can I has multithreading? (Need parallel)}
}
\description{
Create MzObj from directory or file
}
