% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_1_extract_old_way.R
\name{mzml_magic}
\alias{mzml_magic}
\title{MZML Extract Magic: into MZ_OBJ}
\usage{
mzml_magic(input_dir = getwd(), threads = 2)
}
\arguments{
\item{threads}{\cr
Numeric : How many threads?}

\item{input_mzml}{\cr
String : Filename.mzml}
}
\value{
list\link{mz_obj-pos, mz_obj-neg}
}
\description{
MZML->MZ_OBJ for all your MZ_OBJ needs
\itemize{
\item Requires: Bioc::mzR, dplyr, pbapply
}
}
