% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_1_extract.R
\name{mzTime_squashMz}
\alias{mzTime_squashMz}
\title{MzT OBJ Squash mz (simple binning)}
\usage{
mzTime_squashMz(mzT, method = mean, bin_size = 0.005, ignore_nulls = TRUE)
}
\arguments{
\item{mzT}{\cr
mzT Object}

\item{method}{String: Mean? Median? etc?}

\item{bin_size}{\cr
Float: Allowed size of bins}

\item{ignore_nulls}{\cr
Boolean: Ignore null in the calculation?}
}
\value{
\link{mzT-pos, mzT-neg}
}
\description{
Squash the mz element of MzT OBJ.
\itemize{
\item Requires: Dplyr?
}
}
