% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_1_extract.R
\name{mzTime_squashTime}
\alias{mzTime_squashTime}
\title{MzT OBJ Squash time}
\usage{
mzTime_squashTime(mzT, method = mean, ignore_nulls = TRUE)
}
\arguments{
\item{mzT}{\cr
mzT Object}

\item{method}{String: Mean? Median? etc?}

\item{ignore_nulls}{\cr
Boolean: Ignore null in the calculation?}
}
\value{
\link{mzT-pos, mzT-neg}
}
\description{
Squash the time element of MzT OBJ.
\itemize{
\item Requires: Dplyr
}
}
