% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_1_extract_centroid.R
\name{mzT_squashTime}
\alias{mzT_squashTime}
\title{Squash MzT obj (many scans) into a single mz|intensity dataframe}
\usage{
mzT_squashTime(mzT, method = mean, ignore_nulls = T)
}
\arguments{
\item{mzT}{\cr
MzT : MzT object}

\item{method}{\cr
\link{Math} : i.e. (mean, max, median)}

\item{ignore_nulls}{\cr
Boolean: Consider 0 == null, thena ignore them?}
}
\description{
Squash MzT obj (many scans) into a single mz|intensity dataframe
}
